refactor structure for google and github buttons
