+++ src/features/githubButton/GithubAuthButton.tsx
@@ -6,7 +6,7 @@ import GitHub from '@/../public/assets/components/GithubSvgrepoCom';
 export const GithubAuthButton = () => {
   return (
     <button onClick={() => window.location.assign('https://inctagram.work/api/v1/auth/github/login')}>
-      <GitHub />
+      <GitHub height={36} width={36} />
     </button>
   );
 };
