+++ src/features/index.ts
@@ -1,6 +1,8 @@
 export * from './addComment/AddComment';
 export * from './changePhotoProfile/ChangePhoto';
 export * from './generalInfoNavigation/GeneralInfoNavigation';
+export * from './githubButton/GithubAuthButton';
+export * from './googleButton/ui/GoogleAuthButton';
 export * from './localeSwitcher/ui/LocaleSwitcher';
 export * from './localeSwitcher/ui/localeMenuItem/LocaleMenuItem';
 export * from './moreActions/MoreActions';
